requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
et-xmlfile==1.0.1
idna==2.10
jdcal==1.4.1
openpyxl==3.0.4
python-dotenv==0.14.0
requests==2.24.0
selenium==3.141.0
soupsieve==2.0.1
urllib3==1.25.10
